requirements.txt
python-docx==1.1.2
langchain-text-splitters==0.2.4
chromadb==0.5.5
numpy==1.26.4
langchain-ollama==0.1.3
rank-bm25
aiogram